restructured main
